Moved tests/s3.py and utils.py to a separate directory
