+++ spark-testing/spark_utils.py
@@ -1,19 +1,16 @@
 import dcos.config
 import dcos.http
 import dcos.package
-
 import logging
 import os
 import re
+import sdk_security
 import shakedown
 import subprocess
 import urllib
 import urllib.parse
 
-import sdk_security
-
-import s3
-
+import spark_s3
 
 SPARK_SERVICE_ACCOUNT = os.getenv("SPARK_SERVICE_ACCOUNT", "spark-service-acct")
 SPARK_SERVICE_ACCOUNT_SECRET = os.getenv("SPARK_SERVICE_ACCOUNT_SECRET", "spark-service-acct-secret")
@@ -195,10 +192,10 @@ def upload_file(file_path):
         os.environ['S3_BUCKET'],
         os.environ['S3_PREFIX']))
 
-    s3.upload_file(file_path)
+    spark_s3.upload_file(file_path)
 
     basename = os.path.basename(file_path)
-    return s3.http_url(basename)
+    return spark_s3.http_url(basename)
 
 
 def submit_job(app_url, app_args, app_name=SPARK_APP_NAME, args=[], spark_user=SPARK_USER,
@@ -269,7 +266,7 @@ def teardown_spark():
 
 
 def _scala_test_jar_url():
-    return s3.http_url(os.path.basename(os.environ["SCALA_TEST_JAR_PATH"]))
+    return spark_s3.http_url(os.path.basename(os.environ["SCALA_TEST_JAR_PATH"]))
 
 
 def spark_security_session():
