+++ tests/test_kafka.py
@@ -4,7 +4,8 @@ import pytest
 import json
 import shakedown
 
-from tests import s3, utils
+import spark_s3 as s3
+import spark_utils as utils
 
 import sdk_auth
 import sdk_cmd
